+++ tests/pytest/excel/archivos_de_prueba/clases_carrera_vacía.xlsx
