+++ tests/pytest/excel/archivos_de_prueba/clases_nominal.xlsx
