+++ tests/pytest/excel/conftest.py
@@ -1,5 +1,39 @@
+from openpyxl.worksheet.worksheet import Worksheet
+from openpyxl import load_workbook
+from pathlib import Path
+import pytest
+
 from mocks import (
     pytest_configure,
     edificios,
     carreras
 )
+
+archivos_de_prueba = Path(__file__).parent.resolve() / 'archivos_de_prueba'
+
+def pytest_configure(config):
+    # Registrar los markers usados por las fixtures
+    config.addinivalue_line("markers", "archivo: marca para pasar parametros al fixture archivo")
+
+@pytest.fixture
+def archivo(request) -> Path:
+    '''
+    Recibe en el marker "archivo" un nombre de archivo del directorio
+    "archivos_de_prueba".
+
+    :return: El path absoluto del archivo.
+    '''
+    nombre = request.node.get_closest_marker('archivo').args[0]
+    return archivos_de_prueba / nombre
+
+@pytest.fixture
+def primera_hoja_del_archivo(archivo) -> Worksheet:
+    '''
+    Recibe en el marker "archivo" el nombre de un archivo excel del directorio
+    "archivos_de_prueba".
+
+    :return: La primera hoja del archivo.
+    '''
+    libro = load_workbook(archivo)
+    hoja = libro.active
+    return hoja
